Added stamina mechanics
Added the skeleton of some basic stamina mechanics. These mechanics allow the player to run faster at the cost of stamina. As of current, there is no penalty for running out, but that has potential to be added later as additional testing is performed. This may also necessitate a branch of the game with this mechanic absent.