Fix CameraGroups
Scrolling behavior is now handled within the Camera object rather than the Level object, which to me makes more sense.

We've also changed scrolling behavior so that scrolling occurs on scrolling at any time rather than only when the player is touching the ground. This is more bug-prone, but feels a bit better depending. This also cuts down considerably on the complexity of camera functions, making everything more readable.

Camera scrolling behavior is not set in stone here, and open to revision depending.