Added animate module.
Added a module for animation functions. For now only contains the animate_sprite_dict function, which controls animations for dictionaries of frame lists. This is where other functions relevant to animating sprites will live. 

Also includes an if __name__ == "__main__": section for animation testing. This file can be run outside main.py if one wants to test the sprite animations.