Added sprite config utility.
Created a utility script that automates building animation config files in JSON. Compatible with all current code. Builds files by prompting for a list of player states, number of frames in those states, read positions and dimensions. Outputs a JSON file in the same directory.

1. For sprite sheets with static sprite sizes (all frames are the same dimensions), fully automates creation of configuration files for the entire file, for all animation state data provided.
2. For sprite sheets with dynamic sprite sizes (individual frames vary in their dimensions), prompts user to input all sprite position/size data manually. This is very tedious, but still faster than building full JSON files from scratch.

Notes: This file/directory structure may be changed in the future.